Добавляется презентер movieList
